Added logging redirect to evmserver script.
During the update process evmserver.sh update_start is called from
within a yum posttrans and yum's parent process has already been killed as
a part of server shutdown.
This caused the rake task started by evmserver to be killed when
attempting to write to stdout or stderr.
For this reason the redirect was added for rake tasks in evmserver.sh
that are called during an update.

https://bugzilla.redhat.com/show_bug.cgi?id=1239035