Run pre-rotate script even if source command fails
There are some things that the various files sourced
by /etc/default/evm do that SELinux will prevent based
on the context logrotate runs under. Because of this
we need to source the file in order to set our
environment variables and run the pre-rotate script
regardless of the exit status of the source command.

https://bugzilla.redhat.com/show_bug.cgi?id=1265463