Change systemd service KillMode to process
We need the evmserverd process to exit while performing
an update.
To do this we run `systemctl stop evmserverd` which brings
down the main process in the tree managed by systemd.
When the main process exits, systemd sends a sigterm to all
other processes in the tree, killing the update process.

This change will tell systemd to not kill other processes in
the tree after the main one has exited.

See `man systemd.kill` for more info.

https://bugzilla.redhat.com/show_bug.cgi?id=1239035