Removing the reindex script for miq_servers
This table is unlikely to need reindexing as the indexes are not
updated and rows are not removed very often.

The indexes as of this commit on miq_servers are on guid, vm_id,
and zone_id. None of which are likely to change. Because of this,
we are not going to encounter index bloat which means this reindex
is not worth the hourly lock taken on the table.