Record then recreate logical replication metadata across the upgrade
Replication slots are not persisted across pg_upgrade invocations.
I'm not sure for the reason for this, but parts may be because the
new, upgraded, database is create first by initdb and then later
populated with the upgraded data or because the slot state is stored
on the filesystem.
Either way, we need to have the slot in place after upgrade for
pglogical to work properly.

Additionally PostgreSQL 9.5 introduced a concept called replication
origins (https://www.postgresql.org/docs/9.5/static/replication-origins.html)
which allow for replication progress tracking from the consumer.

For pglogical subscriptions to work with 9.5 we need to also create
a replication origin record using the subscription's slot name.

https://bugzilla.redhat.com/show_bug.cgi?id=1391693