Define RAILS_ROOT for the console
We used to define it on the fly if it wasn't already defined, but
now that gems-pending is it's own gem we can't rely on the path of
the files to figure out where RAILS_ROOT should be.

This is causing many of the options to fail.
Defining the constant here will allow the console to work while we
work on fixing the gems-pending code to do something better.