Add trailing slash to cockpit machine urls
Work around for cockpits login page not setting a content-type.
This was fixed back in May, but hasn't made it to rhel yet.

This means that if the machine to connect to contains a . apache
might guess the wrong mime type. We can avoid the issue by adding
a trailing slash to the machine url if one is not present.