Use a login shell for the failover monitor service
As is, the ExecStart command was not working properly.
The issue can be seen with the following trivial example which
will not print anything.

sh -c THING=1 && echo "$THING"

Sourcing /etc/default/evm is done when using a login shell so we
can avoid the issue of how to source the script properly in a one
line command.

https://bugzilla.redhat.com/show_bug.cgi?id=1512985