Use a better command to add /usr/local/bin to the path
Previously when /usr/local/bin was already on the path the command
would return an error status because the initial command was not
successful. With the new command a success status will be returned
in both cases.

This was failing the container builds because `source /etc/default/evm`
was exiting with 1