Remove SSLCipherSuite
As recommended by Red Hat Security, this SSLCipherSuite was
cargo-culted, and it's better to rely upon the default RHEL/CentOS
SSLCipherSuite.  In RHEL8/CentOS8 (which the ManageIQ appliance is built
on), the defaults are better than what's specified here and will be kept
up to date.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1784940