Remove specs and Ruby
While we may want specs in the future, there aren't any now, so a green
PR gives us the illusion of us having coverage which doesn't exist which
is arguably worse.

This commit removes rspec, and after removing rspec, there is no longer
a reason to have Ruby either, so it removes all that too. There's no
code to check style on nor yaml to lint, so that all goes too.