Add bundler aware shell alias for the cli
Similar change as https://github.com/ManageIQ/manageiq-appliance/pull/393

Details copied here:

We were activating a 0.2.0 version first, then trying to load the bundle,
which includes 0.1.2. This fails with:

"You have already activated base64 0.2.0, but your Gemfile requires base64 0.1.2.
Since base64 is a default gem, you can either remove your dependency on it or
try updating to a newer version of bundler that supports base64 as a default gem."

Perhaps the change is that activesupport 7.1 added base64
as a direct dependency.

See:
https://rubygems.org/gems/activesupport/versions/7.1.5.1
vs.
https://rubygems.org/gems/activesupport/versions/7.0.8.7

An alternative is to add another binstub earlier in the $PATH paths.